third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/sandbox-empty-subframe-expected.txt

CONSOLE ERROR: Blocked script execution in 'data:text/html,<script>console.log('Script executed in child frame');</script>' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
 This test passes if the grandchild iframe doesn't log a message.
